requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.20
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.25.2
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
alembic>=1.13.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
openai>=1.6.1